BugB.js - change the animation to fit 32*32 - change size = x:32, y:32 - override the draw function. Now the function will draw the bug double the context size. However, this will not call any parent draw function (e.g. debug draw will not draw the entity borders)
BugC.js
- change size to x:32, y:32

ice.js
- ice will now stop the bug completly instead of slow it down. this was
due to problem with the bug acceleration (maybe the calculation is
wrong). 

playerBase.js
- set health to 100

main.js
- change the debugging method. instead of right-clicking to spawn a bug,
you can press 1 for bugA, 2 for bugB and 3 for bugC. when the umber is
pressed, the bug will spawn at the mouse position.
- fixed a problem when player click "Try Again" button, the upgrades is
still active althought the shop shows that the upgrades level is 1. this
happens due to not clearing the upgreadables array. the old entities in
the array are still alive. this was fixed by setting the array to
'empty' in the start function.

todo.txt
- updated